requirements.txt
flask
transformers
diffusers
torch
Pillow
